refactor(ES-025): cli/setup.ts 分割（setup-ui / setup-fs / setup-context）
539行の setup.ts をコンソール出力・ファイルシステム・プロジェクトコンテキストの
3サブモジュールに分割し、288行まで削減。

Closes #151
Closes #152
