+++ packages/jimmy/src/cli/setup.ts
@@ -1,8 +1,6 @@
-import { execSync, spawn } from "node:child_process";
+import { spawn } from "node:child_process";
 import fs from "node:fs";
-import os from "node:os";
 import path from "node:path";
-import readline from "node:readline";
 import yaml from "js-yaml";
 import { initDb } from "../sessions/registry.js";
 import {
@@ -20,258 +18,9 @@ import {
   TMP_DIR,
 } from "../shared/paths.js";
 import { getPackageVersion } from "../shared/version.js";
-
-const GREEN = "\x1b[32m";
-const YELLOW = "\x1b[33m";
-const RED = "\x1b[31m";
-const DIM = "\x1b[2m";
-const RESET = "\x1b[0m";
-
-function ok(msg: string) {
-  console.log(`  ${GREEN}[ok]${RESET} ${msg}`);
-}
-
-function warn(msg: string) {
-  console.log(`  ${YELLOW}[warn]${RESET} ${msg}`);
-}
-
-function fail(msg: string) {
-  console.log(`  ${RED}[missing]${RESET} ${msg}`);
-}
-
-function info(msg: string) {
-  console.log(`  ${DIM}${msg}${RESET}`);
-}
-
-function prompt(question: string, defaultValue?: string): Promise<string> {
-  const rl = readline.createInterface({ input: process.stdin, output: process.stdout });
-  const suffix = defaultValue ? ` ${DIM}(${defaultValue})${RESET}` : "";
-  return new Promise((resolve) => {
-    rl.question(`  ${question}${suffix}: `, (answer) => {
-      rl.close();
-      resolve(answer.trim() || defaultValue || "");
-    });
-  });
-}
-
-function whichBin(name: string): string | null {
-  try {
-    const cmd = process.platform === "win32" ? "where" : "which";
-    return execSync(`${cmd} ${name}`, { encoding: "utf-8" }).trim().split("\n")[0];
-  } catch {
-    return null;
-  }
-}
-
-function runVersion(bin: string): string | null {
-  try {
-    return execSync(`${bin} --version`, { encoding: "utf-8", timeout: 10000 }).trim();
-  } catch {
-    return null;
-  }
-}
-
-function ensureDir(dir: string): boolean {
-  if (fs.existsSync(dir)) return false;
-  fs.mkdirSync(dir, { recursive: true });
-  return true;
-}
-
-function ensureFile(filePath: string, content: string): boolean {
-  if (fs.existsSync(filePath)) return false;
-  fs.mkdirSync(path.dirname(filePath), { recursive: true });
-  fs.writeFileSync(filePath, content, "utf-8");
-  return true;
-}
-
-/**
- * Apply template placeholder replacements to file content.
- * Only applies to .md and .yaml files.
- */
-function applyTemplateReplacements(content: string, replacements: Record<string, string>): string {
-  let result = content;
-  for (const [placeholder, value] of Object.entries(replacements)) {
-    result = result.replaceAll(placeholder, value);
-  }
-  return result;
-}
-
-/**
- * Recursively copy template directory contents into dest, skipping files that already exist.
- * Applies template placeholder replacements to .md and .yaml files.
- * Returns list of created file paths.
- */
-function copyTemplateDir(srcDir: string, destDir: string, replacements?: Record<string, string>): string[] {
-  const created: string[] = [];
-  if (!fs.existsSync(srcDir)) return created;
-
-  fs.mkdirSync(destDir, { recursive: true });
-
-  const entries = fs.readdirSync(srcDir, { withFileTypes: true });
-  for (const entry of entries) {
-    const srcPath = path.join(srcDir, entry.name);
-    const destPath = path.join(destDir, entry.name);
-
-    if (entry.isDirectory()) {
-      created.push(...copyTemplateDir(srcPath, destPath, replacements));
-    } else if (entry.name === ".gitkeep") {
-    } else if (!fs.existsSync(destPath)) {
-      fs.mkdirSync(path.dirname(destPath), { recursive: true });
-      const ext = path.extname(entry.name).toLowerCase();
-      if (replacements && (ext === ".md" || ext === ".yaml" || ext === ".yml")) {
-        const content = fs.readFileSync(srcPath, "utf-8");
-        fs.writeFileSync(destPath, applyTemplateReplacements(content, replacements), "utf-8");
-      } else {
-        fs.copyFileSync(srcPath, destPath);
-      }
-      created.push(destPath);
-    }
-  }
-  return created;
-}
-
-/**
- * Detect project context by scanning ~/Projects/ for common project indicators
- * and suggest relevant skills the user might want to install.
- */
-function detectProjectContext(portalSlug: string): void {
-  const projectsDir = path.join(os.homedir(), "Projects");
-  if (!fs.existsSync(projectsDir)) return;
-
-  const indicators: { check: (dir: string) => boolean; query: string; label: string }[] = [
-    {
-      check: (dir) => {
-        try {
-          return fs.readdirSync(dir).some((e) => e.endsWith(".xcodeproj"));
-        } catch {
-          return false;
-        }
-      },
-      query: "ios swift xcode",
-      label: "iOS",
-    },
-    {
-      check: (dir) => fs.existsSync(path.join(dir, "Package.swift")),
-      query: "ios swift xcode",
-      label: "iOS/Swift",
-    },
-    {
-      check: (dir) => fs.existsSync(path.join(dir, "Dockerfile")),
-      query: "docker container",
-      label: "Docker",
-    },
-    {
-      check: (dir) => fs.existsSync(path.join(dir, ".github", "workflows")),
-      query: "github actions ci",
-      label: "GitHub Actions",
-    },
-    {
-      check: (dir) => {
-        try {
-          return fs.readdirSync(dir).some((e) => e.startsWith("playwright.config"));
-        } catch {
-          return false;
-        }
-      },
-      query: "playwright testing",
-      label: "Playwright",
-    },
-    {
-      check: (dir) => {
-        const pkgPath = path.join(dir, "package.json");
-        if (!fs.existsSync(pkgPath)) return false;
-        try {
-          const pkg = JSON.parse(fs.readFileSync(pkgPath, "utf-8"));
-          const deps = { ...pkg.dependencies, ...pkg.devDependencies };
-          return deps != null && ("react" in deps || "next" in deps);
-        } catch {
-          return false;
-        }
-      },
-      query: "react nextjs",
-      label: "React",
-    },
-  ];
-
-  const detected = new Map<string, string>(); // label → query
-
-  try {
-    const topLevel = fs.readdirSync(projectsDir, { withFileTypes: true }).filter((e) => e.isDirectory());
-
-    const projectDirs: string[] = [];
-    for (const dir of topLevel) {
-      const dirPath = path.join(projectsDir, dir.name);
-      projectDirs.push(dirPath);
-      // One level deeper for org-style folders (e.g. ~/Projects/Personal/foo)
-      try {
-        const subDirs = fs.readdirSync(dirPath, { withFileTypes: true }).filter((e) => e.isDirectory());
-        for (const sub of subDirs) {
-          projectDirs.push(path.join(dirPath, sub.name));
-        }
-      } catch {
-        // ignore permission errors
-      }
-    }
-
-    for (const projDir of projectDirs) {
-      for (const ind of indicators) {
-        if (detected.has(ind.label)) continue;
-        if (ind.check(projDir)) {
-          detected.set(ind.label, ind.query);
-        }
-      }
-    }
-  } catch {
-    return;
-  }
-
-  if (detected.size > 0) {
-    console.log("");
-    for (const [label, query] of detected) {
-      console.log(
-        `  💡 Detected ${label} projects. Run ${DIM}${portalSlug} skills find ${query}${RESET} to discover relevant skills.`,
-      );
-    }
-  }
-}
-
-const DEFAULT_CONFIG = `jinn:
-  version: "${getPackageVersion()}"
-
-gateway:
-  port: 7777
-  host: "127.0.0.1"
-engines:
-  default: claude
-  claude:
-    bin: claude
-    model: opus
-    effortLevel: medium
-  codex:
-    bin: codex
-    model: gpt-5.4
-connectors: {}
-portal: {}
-logging:
-  file: true
-  stdout: true
-  level: info
-`;
-
-function defaultClaudeMd(portalName: string) {
-  return `# ${portalName} AI Gateway
-
-This is the ${portalName} home directory (~/.jinn).
-${portalName} orchestrates Claude Code and Codex as AI engines.
-`;
-}
-
-function defaultAgentsMd(portalName: string) {
-  return `# ${portalName} Agents
-
-Agents are configured via employees in the org/ directory.
-`;
-}
+import { DEFAULT_CONFIG, defaultAgentsMd, defaultClaudeMd, detectProjectContext } from "./setup-context.js";
+import { applyTemplateReplacements, copyTemplateDir, ensureDir, ensureFile, runVersion, whichBin } from "./setup-fs.js";
+import { DIM, fail, GREEN, info, ok, prompt, RESET, warn, YELLOW } from "./setup-ui.js";
 
 export async function runSetup(opts?: { force?: boolean }): Promise<void> {
   console.log("\nJinn Setup\n");
